refactor: minor renamings
