+++ src/modules/hunter-registry.js
@@ -1,3 +1,7 @@
+// Type-hinting imports
+// eslint-disable-next-line no-unused-vars
+const { Message } = require('discord.js');
+
 const Logger = require('./logger');
